[cli] migrating some functions from the test file to helpers
